Move DoneSignal to utils folder
