+++ profile_bluesky/startup/instrument/devices/lakeshore340.py
@@ -10,19 +10,7 @@ from ophyd import Component, Device, Signal
 from ophyd import EpicsSignal, EpicsSignalRO
 from ophyd import FormattedComponent,PVPositioner
 from ophyd import Kind
-
-class DoneSignal(Signal):
-    def get(self,**kwargs):
-        readback = self.parent.readback.get()
-        setpoint = self.parent.setpoint.get()
-        tolerance = self.parent.tolerance
-
-        if abs(readback-setpoint) <= tolerance:
-            self.put(1)
-        else:
-            self.put(0)
-
-        return self._readback
+from ..utils import DoneSignal
 
 class LS340_LoopBase(PVPositioner):
